requirements.txt
torch>=2.0.0
torchvision>=0.15.0
diffusers>=0.25.0
transformers>=4.35.0
accelerate>=0.24.0
pillow>=10.0.0
safetensors>=0.4.0
flask>=3.0.0
